Break out ReadyToPrintDocuments
